+++ src/internal/create-cli.ts
@@ -16,7 +16,7 @@ export type CliOptions = {
   input: string;
   output: string;
   exportOnly: boolean;
-  skipReview: boolean;
+  reviewNames: boolean;
   skipRelatedEntities: boolean;
   entities: Array<Entity>;
   objectsDir: string;
@@ -84,7 +84,7 @@ export function createCLI(argv = process.argv) {
     .option("-i, --input <path>", "*.ts/*.d.ts file")
     .option("-o, --output <dir>", "output root directory", DEFAULT_ROOT_DIR)
     .option("--export-only", "only extract exported interfaces/types", false)
-    .option("--skip-review", "skip review names after selection", false)
+    .option("--review-names", "review names after selection", false)
     .option("--skip-related-entities", "skip generating related entities (views, navItems)", false)
     .option("-e, --entities [entities...]", `can specify single or multiple entities among: ${styleText("yellow", "(object | view | navItem)")}`)
     .option("--objects-dir <dir>", "output object directory", DEFAULT_OBJECTS_DIR)